+++ dev.js
@@ -7,6 +7,7 @@ import LineString from 'ol/geom/LineString';
 import Modify from 'ol/interaction/Modify';
 import {
   MaplibreLayer,
+  MaplibreStyleLayer,
   MapsetAPI,
   MapsetLayer,
   MocoLayer,
@@ -23,36 +24,52 @@ const mapsUrl = 'https://maps.geops.io';
 
 const baseLayer = new MaplibreLayer({
   apiKey: window.apiKey,
-  style: 'travic_v2',
-  // style: 'de.rvf_moco',
+  // style: 'tralis_munich_schematic_v3',
+  style: 'de.rvf_moco',
+  // url: 'https://maps.geops.io/styles/tralis_munich_schematic_v3/style.json?key=5cc87b12d7c5370001c1d655112ec5c21e0f441792cfc2fafe3e7a1e',
 });
 
 const mapsetLayer = new MapsetLayer({
+  // apiKey: window.apiKey,
+  // tenants: ['rvf'],
   apiKey: window.apiKey,
-  tenants: ['rvf'],
+  tags: ['mobility-portal-sob'],
+  tenants: ['geopsmarketing'],
+});
+
+const lnpLayer = new MaplibreStyleLayer({
+  isQueryable: true,
+  // hideInLegend: true,
+  layersFilter: (l) => {
+    return l.source === 'network_plans';
+  },
+  maplibreLayer: baseLayer,
 });
 
 const realtimeLayer = new RealtimeLayer({
   apiKey: window.apiKey,
   // apiKey: '5cc87b12d7c5370001c1d655112ec5c21e0f441792cfc2fafe3e7a1e', // sbm
-  url: 'wss://api.geops.io/tracker-ws/v1/', // prod
+  // url: 'wss://api.geops.io/tracker-ws/v1/', // prod
   // url: 'wss://api.geops.io/realtime-ws/v1/', // sbm
+  url: 'https://api.geops.io/tracker-http/v1/', // rvf
   styleOptions: {
     // useHeadingStyle: true,
     // useDelayStyle: true,
   },
   // style: realtimeByMotStyle,
   // tenant: 'sbm',
-  // tenant: 'trenord',
-  // bboxParameters: {
-  //   line_tags: 'RVF',
-  // },
+  bboxParameters: {
+    line_tags: 'RVF',
+  },
+  visible: true,
 });
 
 const mocoLayer = new MocoLayer({
   apiKey: window.apiKey,
-  // tenant: 'rvf',
+  tenant: 'rvf',
   maplibreLayer: baseLayer,
+  lnpLayer: lnpLayer,
+  loadByZoom: true,
   // publicAt: new Date(),
   // url: mocoUrl,
 });
@@ -60,16 +77,18 @@ const mocoLayer = new MocoLayer({
 const map = new Map({
   layers: [
     baseLayer,
+    lnpLayer,
+    mocoLayer,
+    // mapsetLayer,
     realtimeLayer,
-    // mocoLayer,
-    // mapsetLayer
   ],
   target: 'map',
   view: new View({
-    // center: [872814.6006106276, 6106276.43], // rvf
-    center: [1022769, 5698188], // trenord
+    center: [872814.6006106276, 6106276.43], // rvf
+    // center: [1022769, 5698188], // trenord
     // center: [1286668, 6130216], // sbm
-    zoom: 16,
+    // center: [2383522, 1674321], // sbm schematic
+    zoom: 13,
   }),
 });
 
@@ -106,7 +125,18 @@ map.on('moveend', () => {
   const zoom = map.getView().getZoom();
   console.log('Current zoom level:', zoom);
 });
-
+const toggleLayerVisibilityButton = document?.getElementById(
+  'toggle-visiblity-button',
+);
+toggleLayerVisibilityButton.addEventListener('click', () => {
+  if (lnpLayer.getVisible()) {
+    lnpLayer.setVisible(false);
+    toggleLayerVisibilityButton.textContent = 'Show LNP Layer';
+  } else {
+    lnpLayer.setVisible(true);
+    toggleLayerVisibilityButton.textContent = 'Hide LNP Layer';
+  }
+});
 // const urlInput = document?.getElementById('url-input');
 // const setUrlButton = document?.getElementById('set-url-button');
 // setUrlButton.addEventListener('click', () => {