+++ src/common/utils/realtimeStyleUtils.ts
@@ -194,11 +194,10 @@ export const getTextSize = (
   }
   ctx.font = font;
   let newText = ctx.measureText(text);
-
   const maxiter = 25;
   let i = 0;
 
-  while (newText.width > markerSize - 6 && i < maxiter) {
+  while (newText.width > markerSize && i < maxiter) {
     const previousFontSize = fontSize;
     // eslint-disable-next-line no-param-reassign
     fontSize -= 0.5;
@@ -217,11 +216,8 @@ export const getDelayColor = (
   if (cancelled) {
     return isDelayText ? '#ff0000' : '#a0a0a0'; // red or gray
   }
-  if (!delayInMs) {
-    return '';
-  }
-  if (delayInMs === null) {
-    return '#a0a0a0'; // grey { r: 160, g: 160, b: 160, s: '160,160,160' };
+  if (delayInMs === null || delayInMs === undefined) {
+    return ''; // grey { r: 160, g: 160, b: 160, s: '160,160,160' };
   }
   if (delayInMs >= 3600000) {
     return '#ed004c'; // pink { r: 237, g: 0, b: 76, s: '237,0,76' };